Move monitor functionality out of the main window example
